tools/graphrag-benchmark/pyproject.toml
[project]
name = "graphrag-benchmark"
version = "0.1.0"
description = "Selective GraphRAG benchmark tooling for EduAssist Platform."
requires-python = ">=3.12"
dependencies = [
  "graphrag==3.0.6",
  "PyYAML>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graphrag_benchmark"]
